% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.nnet}
\alias{train.nnet}
\title{train.nnet}
\usage{
train.nnet(formula, data, weights, ..., subset, na.action,
  contrasts = NULL)
}
\description{
train.nnet
}
\examples{

data("iris")

n <- seq_len(nrow(iris))
.sample <- sample(n, length(n) * 0.75)
data.train <- iris[.sample,]
data.test <- iris[.sample,]

real <- data.test$Species

modelo.nn <- train.nnet(Species~., data.train, size = 20)
modelo.nn
prob <- predict(modelo.nn, data.test, type = "prob")
prob
prediccion <- predict(modelo.nn, data.test, type = "class")
prediccion
table(real, prediccion)

}
