% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.svm}
\alias{train.svm}
\title{train.svm}
\usage{
train.svm(formula, data, ..., subset, na.action = na.omit,
  scale = TRUE)
}
\description{
train.svm
}
\examples{

data("iris")

n <- seq_len(nrow(iris))
.sample <- sample(n, length(n) * 0.75)
data.train <- iris[.sample,]
data.test <- iris[.sample,]

real <- data.test$Species

modelo.svm <- train.svm(Species~., data.train)
modelo.svm
prob <- predict(modelo.svm, data.test , type = "prob")
prob
prediccion <- predict(modelo.svm, data.test , type = "class")
prediccion
table(real, prediccion)

}
