% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{varplot}
\alias{varplot}
\title{Plotting prmdt ada models}
\usage{
varplot(x, ...)
}
\arguments{
\item{x}{A ada prmdt model}

\item{...}{optional arguments to print o format method}
}
\description{
Plotting prmdt ada models
}
