% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.xgboost}
\alias{train.xgboost}
\title{train.xgboost}
\usage{
train.xgboost(formula, data, nrounds, watchlist = list(), obj = NULL,
  feval = NULL, verbose = 1, print_every_n = 1L,
  early_stopping_rounds = NULL, maximize = NULL, save_period = NULL,
  save_name = "xgboost.model", xgb_model = NULL, callbacks = list(),
  early_stop_round = 10, eval_metric = "mlogloss",
  extra_params = NULL, booster = "gbtree", objective = NULL,
  eta = 0.3, gamma = 0, max_depth = 6, min_child_weight = 1,
  subsample = 1, colsample_bytree = 1, ...)
}
\description{
train.xgboost
}
\examples{

data("iris")

n <- seq_len(nrow(iris))
.sample <- sample(n, length(n) * 0.75)
data.train <- iris[.sample,]
data.test <- iris[.sample,]

real <- data.test$Species

modelo.xg <- train.xgboost(Species~., data.train, nrounds = 79, maximize = FALSE)
modelo.xg
prob <- predict(modelo.xg, data.test, type = "prob")
prob
prediccion <- predict(modelo.xg, data.test, type = "class")
prediccion
table(real, prediccion)

}
